Send extractor results via database
