+++ flashcard_model/celery_consumer.py
@@ -1,6 +1,7 @@
 from celery import Celery
 
-from flashcard_model.flashcard_model.chat_gpt_model import generate_chat_gpt_questions
+from flashcard_model.chat_gpt_model import generate_chat_gpt_questions
+from flashcard_model.database_functions import load_extracted_pages
 
 app = Celery('flashcard_model')
 app.conf.task_default_queue = "flashcard_model"
@@ -10,5 +11,6 @@ app.conf.accept_content = ["json", "pickle"]
 
 @app.task(name="generate_flashcards")
 def generate_flashcard(extraction_result):
-    generate_chat_gpt_questions(extraction_result["document_id"], extraction_result["extracted_pages"])
+    extracted_pages = load_extracted_pages(extraction_result["document_id"])
+    generate_chat_gpt_questions(extraction_result["document_id"], extracted_pages)
     print(f"Document {extraction_result['document_id']} flashcard generation finished")
