+++ flashcard_model/requirements.txt
@@ -1,2 +1,4 @@
 celery==5.2.7
 redis==4.5.4
+Pillow==9.4.0
+pymongo==4.3.3
